refactor: deduplicate parse_type_params and apply_subst, fix innermost-class bug
- Extract parse_type_params_from_decl to node_ext.rs (shared, uses split_depth0_commas)
- lookup.rs and resolution.rs now delegate to the shared function
- resolution.rs apply_subst now delegates to lookup::apply_type_subst
- Fix find_containing_class_name: max_by_key → min_by_key (innermost, not outermost)
- Add ENUM and OBJECT to class kind filter in resolution.rs
